remove widgets folder and rename icons file
