+++ mobile/lib/custom/lists/close_friends_list.dart
