+++ mobile/lib/custom/ride_countdown.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_countdown_timer/index.dart';
 
 import 'package:mobile/colors.dart';
-import 'package:mobile/blip_icons_icons.dart';
+import 'package:mobile/icons.dart';
 import 'package:mobile/fonts.dart';
 
 class RideCountDown extends StatefulWidget {
