+++ mobile/lib/screens/home/rider_home.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
-import 'package:mobile/blip_icons_icons.dart';
+import 'package:mobile/custom/lists/close_friends_list.dart';
+import 'package:mobile/custom/lists/ride_offer_list.dart';
+import 'package:mobile/icons.dart';
 import 'package:mobile/custom/toggle_to_driver.dart';
-import 'package:mobile/custom/RideCountDown.dart';
+import 'package:mobile/custom/ride_countdown.dart';
 import 'package:mobile/models/friend.dart';
 import 'package:mobile/models/ride_offer.dart';
 import 'package:mobile/utils/widget_functions.dart';
-import 'package:mobile/widgets/close_friends_list.dart';
 import 'package:mobile/fonts.dart';
-import 'package:mobile/widgets/ride_offer_list.dart';
 import '../../colors.dart';
 
 class RiderHomeScreen extends StatefulWidget {
